Initial commit: MITgcm code and some notebooks
Added a local copy of the MITgcm, because my package doesn't work out of the box with the latest version on Github (why?). Includes an /experiment/ folder where I have my modified configuration files in /input/ and /code/. Input files are generated by /notebooks/gendata.ipynb, and local copies are sent directly to the /input/ folder to keep track of how each was generated.